+++ R/guerrero.R
@@ -1,15 +1,3 @@
-# #' @param lambda Box-cox transformation parameter
-# #' @param x A vector
-# #' @param .period The length of each subseries (usually the length of seasonal period)
-#' @importFrom stats sd
-lambda_coef_var <- function(lambda, x, .period = 2) {
-  if(all(x == x[1])) return(1)
-  x <- split(x, (seq_along(x)-1)%/%.period)
-  mu_h <- map_dbl(x, mean, na.rm = TRUE)
-  sig_h <- map_dbl(x, sd, na.rm = TRUE)
-  rat <- sig_h / mu_h ^ (1 - lambda)
-  sd(rat, na.rm = TRUE) / mean(rat, na.rm = TRUE)
-}
 
 #' Guerrero's method for Box Cox lambda selection
 #'
@@ -40,8 +28,32 @@ lambda_coef_var <- function(lambda, x, .period = 2) {
 #'
 #' @export
 guerrero <- function(x, lower = -0.9, upper = 2, .period = 2L) {
+  if(all(x == x[1])) return(1)
+
+  # Prepare seasonal subseries
+  n_obs <- length(x)
+  n_subseries <- floor(n_obs/.period)
+  x <- matrix(
+    x[(n_obs - floor(n_subseries * .period) + 1):n_obs],
+    nrow = .period, ncol = n_subseries
+  )
+
   c(lambda_guerrero = optimise(
     lambda_coef_var, c(lower, upper), x = x,
     .period = max(.period, 2)
   )$minimum)
+}
+
+# #' @param lambda Box-cox transformation parameter
+# #' @param x A vector
+# #' @param .period The length of each subseries (usually the length of seasonal period)
+#' @importFrom stats sd
+lambda_coef_var <- function(lambda, x, .period = 2) {
+  # Mean of each subseries
+  mu_h <- apply(x, 2, mean, na.rm = TRUE)
+  # Standard deviation of each subseries
+  sig_h <- apply(x, 2, sd, na.rm = TRUE)
+  # Compute fitness metric
+  rat <- sig_h / mu_h ^ (1 - lambda)
+  sd(rat, na.rm = TRUE) / mean(rat, na.rm = TRUE)
 }