+++ NEWS.md
@@ -1,5 +1,10 @@
 # feasts (development version)
 
+## Bug fixes
+
+* Fixed `generate(<STL>)` returning `$.sim` as a `num [1:n(1d)]` instead of 
+  `num [1:72]` (fable/#336).
+
 # feasts 0.2.2
 
 Minor release for compatibility with an upcoming ggplot2 release. This release