+++ NEWS.md
@@ -1,4 +1,4 @@
-# feasts (development version)
+# feasts 0.3.0
 
 ## New features